Convert EditItemForm to tsx and install eslint.
